+++ config/environment.rb
@@ -3,3 +3,4 @@ require_relative 'application'
 
 # Initialize the Rails application.
 Rails.application.initialize!
+
