+++ package.json
@@ -4,6 +4,7 @@
   "dependencies": {
     "@mapbox/mapbox-gl-geocoder": "^4.5.0",
     "@rails/webpacker": "^4.0.7",
+    "@shopify/draggable": "^1.0.0-beta.8",
     "bootstrap": "^4.3.1",
     "jquery": "^3.4.1",
     "mapbox-gl": "^1.5.0",
